requirements.txt
Django==2.2.8
django-admin-daterange-filter==0.1.3
django-admin-rangefilter==0.5.1
django-ajax-selects==1.7.1
django-import-export==1.2.0
django-mathfilters==0.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.1
pytz==2019.3
PyYAML==5.1.2
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0
